refactor(server): move shared function handlers to controllers/shared, one per file
