+++ packages/server/lib/controllers/v1/integrations/providerConfigKey/functions/getCode.ts
@@ -4,7 +4,7 @@ import { zodErrorToHTTP } from '@nangohq/utils';
 
 import { envSchema, functionTypeSchema, providerConfigKeySchema } from '../../../../../helpers/validation.js';
 import { asyncWrapper } from '../../../../../utils/asyncWrapper.js';
-import { handleGetFunctionCode } from './helpers.js';
+import { handleGetFunctionCode } from '../../../../shared/integrations/functions/getCode.js';
 
 import type { GetFunctionCode } from '@nangohq/types';
