+++ packages/server/lib/controllers/v1/integrations/providerConfigKey/functions/getFunction.ts
@@ -4,7 +4,7 @@ import { zodErrorToHTTP } from '@nangohq/utils';
 
 import { envSchema, functionTypeSchema, providerConfigKeySchema } from '../../../../../helpers/validation.js';
 import { asyncWrapper } from '../../../../../utils/asyncWrapper.js';
-import { handleGetIntegrationFunction } from './helpers.js';
+import { handleGetIntegrationFunction } from '../../../../shared/integrations/functions/getFunction.js';
 
 import type { GetIntegrationFunction } from '@nangohq/types';
