+++ packages/server/lib/controllers/v1/integrations/providerConfigKey/functions/getFunctions.ts
@@ -4,8 +4,8 @@ import { zodErrorToHTTP } from '@nangohq/utils';
 
 import { envSchema, functionListQueryFields } from '../../../../../helpers/validation.js';
 import { asyncWrapper } from '../../../../../utils/asyncWrapper.js';
+import { handleListIntegrationFunctions } from '../../../../shared/integrations/functions/listFunctions.js';
 import { validationParams } from '../getIntegration.js';
-import { handleListIntegrationFunctions } from './helpers.js';
 
 import type { GetIntegrationFunctions } from '@nangohq/types';
